Move Linux generate source support
Separate out the generate source support on the Linux backend into
linux_generated.go
